refactor: moved some of functions to a new module datadir.py
