+++ src/bonbast/main.py
@@ -11,6 +11,7 @@ from rich.console import Console, ConsoleRenderable
 from rich.table import Table
 
 from server import get_token_from_main_page, get_prices_from_api
+from datadir import save_token, get_token, delete_token
 
 
 price_formatter = "{:,}"
@@ -83,76 +84,6 @@ class Gold:
     def __init__(self, name: str, price: float):
         self.name = name
         self.price = price
-
-
-
-
-
-
-
-
-def get_datadir() -> pathlib.Path:
-    """
-    Returns a parent directory path
-    where persistent application data can be stored.
-
-    # linux: ~/.local/share
-    # macOS: ~/Library/Application Support
-    # windows: C:/Users/<USER>/AppData/Roaming
-    """
-
-    home = pathlib.Path.home()
-
-    if sys.platform == "win32":
-        return home / "AppData/Roaming"
-    elif sys.platform == "linux":
-        return home / ".local/share"
-    elif sys.platform == "darwin":
-        return home / "Library/Application Support"
-
-
-def get_app_directory() -> pathlib.Path:
-    """
-    Returns a directory path
-    where application data can be stored.
-    """
-
-    return get_datadir() / "bonbast"
-
-
-# save token to use later
-def save_token(token: str, date: datetime) -> None:
-    app_dir = get_app_directory()
-
-    try:
-        app_dir.mkdir(parents=True, exist_ok=True)
-    except FileExistsError:
-        pass
-
-    with open(app_dir / 'token.data', 'w') as f:
-        f.write(f'{token}\n{date.isoformat()}')
-
-
-# retrieve token from storage
-def get_token() -> Tuple[Optional[str], Optional[datetime]]:
-    app_dir = get_app_directory()
-
-    try:
-        with open(app_dir / 'token.data', 'r') as f:
-            token, date = f.read().splitlines()
-            return token, datetime.fromisoformat(date)
-    except FileNotFoundError:
-        return None, None
-
-
-# delete saved token
-def delete_token():
-    app_dir = get_app_directory()
-
-    try:
-        os.remove(app_dir / 'token.data')
-    except FileNotFoundError:
-        pass
 
 
 def get_prices():
